More cleanup of the WorldMapStore
