+++ src/index.jsx
@@ -1,8 +1,23 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
+import connect from './connect';
 import { WorldMapStore } from './player-store';
+import { PlayerModel } from './player-model';
+
+const batchEvents = connect();
+const ps = new PlayerModel();
+
+const playerData = batchEvents.map((playerbatch) => {
+  playerbatch.forEach(player => {
+    ps.set(player.id, player);
+  });
+  return {
+    players: ps.players,
+    playerGrid: ps.playerGrid,
+  };
+});
 
 ReactDOM.render(
-  <WorldMapStore />, document.querySelector('#app')
+  <WorldMapStore batchEvents={playerData} />, document.querySelector('#app')
 );
